pyproject.toml
[tool.poetry]
name = "carsimautomation"
version = "0.1.0"
description = ""
authors = ["melonedo <44501064+melonedo@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
pywinauto = "^0.6.8"
matlabengine = "~9.14"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
